MultiProviderをmainで作成
MainページとしてHome,Talk,Profileは一体なのでMultiProviderとして4つのモデルを定義した。
