+++ lib/main.dart
@@ -1,12 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:provider/provider.dart';
 
-import 'package:chat_flutter/di.dart';
 import 'package:chat_flutter/ui/pages/add_friend.dart';
 import 'package:chat_flutter/ui/pages/create_group.dart';
 import 'package:chat_flutter/ui/pages/main/main.dart';
 import 'package:chat_flutter/ui/pages/main/profile/profile_edit.dart';
-import 'package:chat_flutter/ui/pages/main/profile/profile.dart';
 import 'package:chat_flutter/ui/pages/room.dart';
 import 'package:chat_flutter/ui/pages/sign_in.dart';
 import 'package:chat_flutter/ui/pages/sign_up.dart';
@@ -33,8 +30,7 @@ class MyApp extends StatelessWidget {
         '/signUpPage': (context) => SignUpPage(),
         '/signInPage': (context) => SignInPage(),
         '/roomPage': (context) => RoomPage(),
-        '/profilePage': (context) => ProfilePage.wrapped(),
-        '/profileEditPage': (context) => ProfileEditPage(),
+        '/profileEditPage': (context) => ProfileEditPage.wrapped(),
         '/addFriendPage': (context) => AddFriendPage(),
         '/createGroupPage': (context) => CreateGroupPage(),
       },
