+++ lib/ui/pages/main/main.dart
@@ -2,9 +2,12 @@ import 'package:chat_flutter/ui/molecules/home/app_bar.dart';
 import 'package:chat_flutter/ui/molecules/profile/app_bar.dart';
 import 'package:chat_flutter/ui/molecules/talk/app_bar.dart';
 import 'package:chat_flutter/ui/pages/main/home/home.dart';
+import 'package:chat_flutter/ui/pages/main/home/home_controller.dart';
 import 'package:chat_flutter/ui/pages/main/main_controller.dart';
 import 'package:chat_flutter/ui/pages/main/profile/profile.dart';
+import 'package:chat_flutter/ui/pages/main/profile/profile_controller.dart';
 import 'package:chat_flutter/ui/pages/main/talk/talk.dart';
+import 'package:chat_flutter/ui/pages/main/talk/talk_controller.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
@@ -12,8 +15,21 @@ class MainPage extends StatelessWidget {
   MainPage._({Key key}) : super(key:key);
 
   static Widget wrapped(){
-    return new ChangeNotifierProvider<MainController>(
-      create: (_) => MainController(),
+    return MultiProvider(
+      providers:[
+        ChangeNotifierProvider<MainController>(
+          create: (_) => MainController(),
+        ),
+        ChangeNotifierProvider<HomeController>(
+          create: (_) => HomeController(),
+        ),
+        ChangeNotifierProvider<TalkController>(
+          create: (_) => TalkController(),
+        ),
+        ChangeNotifierProvider<ProfileController>(
+          create: (_) => ProfileController(),
+        ),
+      ],
       child: MainPage._(),
     );
   }
@@ -34,9 +50,9 @@ class MainPage extends StatelessWidget {
     ];
 
     final List pages = [
-      HomePage.wrapped(),
-      TalkPage.wrapped(),
-      ProfilePage.wrapped(),
+      HomePage(),
+      TalkPage(),
+      ProfilePage(),
     ];
 
     return Scaffold(
